+++ model/NoticeMessage.js
@@ -10,11 +10,14 @@ module.exports = class NoticeMessage extends Base {
     static getConstants () {
         return {
             RULES: super.RULES.concat([
-                ['notice', 'relation']
+                [['notice', 'recipients'], 'relation']
             ]),
             BEHAVIORS: {
                 relationChange: require('areto/behavior/RelationChangeBehavior'),
                 timestamp: require('areto/behavior/TimestampBehavior')
+            },
+            ATTR_LABELS: {
+                'recipients': 'Direct recipients'
             }
         };
     }
@@ -27,6 +30,11 @@ module.exports = class NoticeMessage extends Base {
     relPopupNotifications () {
         const Class = this.getClass('model/PopupNotification');
         return this.hasMany(Class, 'message', this.PK);
+    }
+
+    relRecipients () {
+        const Class = this.getClass('model/User');
+        return this.hasMany(Class, Class.PK, 'recipients').viaArray();
     }
 };
 module.exports.init(module);