+++ module/rbac/view/model/meta-target/_form.js
@@ -11,15 +11,15 @@ module.exports = class MetaTargetForm extends Base {
         const model = this.data.model;
         const rbac = model.getRbac();
         const metaClass = rbac.baseMeta.getClass(model.get('class'));
-        const navSection = rbac.navMeta.getSection(model.get('navSection'));
+        const section = rbac.navMeta.getSection(model.get('section'));
         return {
             types: this.controller.getMapSelectItems(rbac.VALUE_LABELS.targets),
             view: metaClass ? metaClass.getView(model.get('view')) : null,
             attr: metaClass ? metaClass.getAttr(model.get('attr')) : null,
             state: metaClass ? metaClass.getState(model.get('state')) : null,
             transition: metaClass ? metaClass.getTransition(model.get('transition')) : null,
-            navNode: navSection ? navSection.getNode(model.get('navNode')) : null,
-            navSection,
+            node: section ? section.getNode(model.get('node')) : null,
+            section,
             metaClass
         };
     }