+++ HISTORY.md
@@ -1,3 +1,10 @@
+## 1.8.0
+
+* model/Notice
+    - move methods validator to parent class 
+* module/rbac/controller/ItemController
+    - fix query creation
+
 ## 1.7.0
 
 * module/rbac/model/AssignmentRule